DRY out some code
